\name{as.data.frame-methods}
\docType{methods}
\alias{as.data.frame-methods}
\alias{as.data.frame,ANY-method}
\alias{as.data.frame,BMLgrid-method}
\title{Methods for Function \code{as.data.frame}}
\description{
  Returns the BMLgrid object as a \code{data.frame}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
  %%  ~~describe this method here~~
}

\item{\code{signature(x = "BMLgrid")}}{
  %%  ~~describe this method here~~
  This function returns the BMLgrid object as a data frame, where each
  row represents a car and the three attributes are the color of the
  car, the column of the grid for the car and the row of the grid for
  that car.

}
}}
\keyword{methods}
\keyword{other possible keyword(s)}
