\name{bluecars-methods}
\docType{methods}
\alias{bluecars-methods}
\alias{bluecars,BMLgrid-method}
\title{get \code{bluecars}}
\description{
  This method returns all the blue cars in the \code{BMLgrid}.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "BMLgrid")}}{
  %%  ~~describe this method here~~
  This method returns all the blue cars in the \code{BMLgrid}.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
