\name{getdim-methods}
\docType{methods}
\alias{getdim-methods}
\alias{getdim,BMLgrid-method}
\title{Methods for Function \code{getdim}}
\description{
 Get the dimension of \code{BMLgrid}
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "BMLgrid")}}{
  %%  ~~describe this method here~~
  Get dimension of the \code{BMLgrid}
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
